Fixes unique values for identical Words in wd_map
This fixes a major issue in the wd_map symbol table where we weren't
checking to see if a given representation already had a place in the
table, instead each encountered W was given a unique entry, even if an
identical value already existed in the table.

Here I've used another map to do the reverse lookup, which will come in
handy when we add metaprogramming features to the block item type. The
implementation may change however.