Rename internal String() to Print().
I originally thought it would be nice to use Go's stringer interface,
but as it turns out I want to use some of the fmt package which also
uses it. This causes infinite recusion when printing some types.

Secondly, String() was too general purpose, and by renaming it Print()
it's more apparent that it's intended mostly for the print operation and
isn't actually a generic stringification method.

This also fixes an issue where large Ns were truncated by print due to
strconv.Itoa requiring int instead of int64 as its parameter.