print is only for T and say is higher level
`print` as spec'd was to only work on T items, there was (and yet may be)
a `refl` op to do a debug-worthy inspection of all item types. During
development it become useful to use `print` on different item types, but
this lead to an uncomfortable middle ground. This intends to begin
aleviating that awkwardness.