Current impl of blacklight is a proof-of-concept.
I've written a lot about blacklight in the present tense to help me to focus on keeping it in a working state with an aim towards long term sustainability. However, for visitors seeing it for the first time I don't want to give them the wrong impression of its current state. This is also a good time to think about doing overhauls of the current implementation to bring it closer in line with my vision for it now that I've proven to myself that the compiler and VM are feasible and workable for this subset of features.