Pass mutex by reference
The current version of Golang seems to support this without problem.

I've tested everything and it seems to work just fine.

I'm quite certain that it used to refuse to compile because it no longer
matched the interface.

I still suspect something is screwy because now some types pass by
reference and some pass by value.