Split out decompile and analysis ops
Previously, I implemented a kind of verbose decompilation testing tool
while I was trying to figure out a good way to build a full decompiler.

I've added the capability to all datatypes to be able to dump the
appropriate original code to decompile itself, so it turns out for basic
use cases that actually is effectively decompilation back into source.

The caveat is that there's not currently a good way to handle nested
containers, but this can be handled pretty easily in the future.

It also does not handle system stacks and possibly others. Futher
testing required.

So now I've kept the verbose debugging tool but also implemented a basic
decompiler to produce something very close to the original source.

This is possible because all of the information is already there aside
from whitespace repitition and differentiation.