Moved depth checks into $Eject from Drop
Basically this code caused a stack underflow:

~~~blacklight
( $drop )
~~~

It happened because Eject didn't clean up after itself like Drop had
recently been enabled to do.

There was no sane reason to make Eject work differently, so I rolled all
the code into one function.

The downside of this is that there could be an additional allocation in
Eject which Drop ends up ignoring. Luckily, these functions are not
expected to be churned heavily.