Reworking the VM.debug output
Hides the complete disassembly because without an easy way to get the
offset inside it just displays the same program you're already familiar
with.

The infer function assumes the current byte is an opword and interprets
it accordingly. This is dumb since it might be something else. I broke
it into two pieces so it could be used genericly, but it will eventually
just be a function to look up the string name of an opcode.

Likely the next steps will be:
- store the current and previous complete
  bytecode instructions in the VMstate for easy reference
- use the offset to determine what should be disassembled