Don't panic, return error for wrong arg
This is a kind of WIP commit.

It runs but there's some odd behvaiour documented in vm.go's handle()
function.

Basically, I'm seeing handle() run twice when a nested block encounters
a TypeAssertionError. And as a result, the displayed error is correct
but the VM state is wrong (because the error was caused by a different
VM).

I'm sure it's something obvious that I'll figure out tomorrow.

In any case, this begins to bring runtime errors in-house and provide
better context for them, even if the sourcemap isn't available.